requirement.txt
wikipedia==1.4.0
Mastodon.py==1.3.1
